Rewrite solve_odgp_internal to return iterator
And does not generate large `Vec`s before returning iterators

With this commit, `gridsynth` no longer consumes materialized lists of grid points.
Rather all lists are iterators. Lists are consumed one element at a time, with
early exit when a point is accepted.

Previously synthesizing RZ(pi/4) would not terminate in any realistic amount of time.
With this commit, it completes in 10s or 100s of milliseconds.