Store computed error in gridsynth result
Previously a bool is stored, signifying whether the computed error is too large.

This PR preserves this bool, but also stores the error (in opnorm) itself.

If error computation is requested, the command line program prints this error before the gate string.

Variables and functions with the string "check" have been replaced with "compute_error".