Fix stack overflow when using low precision
* We use the same value for working precision and target precision. This is not
  needed. When the number of digits is 12, certain input (eg. theta == 1.57) causes the program
  to crash with a stack overflow.

This PR implements ensuring the minimum number of decimal digits of precision is 16.

(In fact, there is probably a better choice for working precision and target precision in general.)

In looking for the bug, I also found crashing from numerical instability in the quadratic equation
solver. This PR also improves the stability of this calculation.